Implement rsync deployment
It reads the destination from a git config: hooks.wwwpath. If the
incoming push is on the master branch, it'll check it out to a temporary
directory (using mktemp) and rsync the src/ directory to the wwwpath.

It'll abort if it fails to checkout the files to the temporary directory
or if hooks.wwwpath is undefined.